refactor: standardize import paths and add Certification entity
