+++ src/contexts/CoreContext/domain/entities/Client.ts
@@ -1,6 +1,6 @@
-import { UniqueEntityID } from "@/contexts/Shared/Domain/UniqueEntityID";
+import { Entity } from "@/contexts/Shared/domain/Entity";
+import { UniqueEntityID } from "@/contexts/Shared/domain/UniqueEntityID";
 import { UserId } from "../valueObjects/UserId";
-import { Entity } from "@/contexts/Shared/Domain/Entity";
 
 interface ClientProps {
   userId: UserId;
