+++ src/contexts/CoreContext/domain/entities/Experience.ts
@@ -1,5 +1,5 @@
-import { Entity } from "@/contexts/Shared/Domain/Entity";
-import { UniqueEntityID } from "@/contexts/Shared/Domain/UniqueEntityID";
+import { Entity } from "@/contexts/Shared/domain/Entity";
+import { UniqueEntityID } from "@/contexts/Shared/domain/UniqueEntityID";
 
 interface ExperienceProps {
   position: string;
