+++ src/contexts/CoreContext/domain/interfaces/repositories/IUserRepository.ts
@@ -1,4 +1,4 @@
-import { IRepository } from "@/contexts/Shared/Domain/repository/IRepository";
+import { IRepository } from "@/contexts/Shared/domain/repository/IRepository";
 import { User } from "../../aggregates/User";
 
 export interface IUserRepository extends IRepository<User> {
