+++ src/contexts/CoreContext/domain/interfaces/services/IFreelancerService.ts
@@ -1,4 +1,4 @@
-import { IService } from "@/contexts/Shared/Domain/service/IService";
+import { IService } from "@/contexts/Shared/domain/service/IService";
 import { Freelancer } from "../../entities/Freelancer";
 
 export interface IFreelancerService extends IService<Freelancer> {
