+++ src/contexts/CoreContext/domain/interfaces/services/IUserService.ts
@@ -1,4 +1,4 @@
-import { IService } from "@/contexts/Shared/Domain/service/IService";
+import { IService } from "@/contexts/Shared/domain/service/IService";
 import { User } from "../../aggregates/User";
 export interface IUserService extends IService<User> {
   delete(id: string): Promise<string | void>;
