+++ src/contexts/CoreContext/domain/valueObjects/Language.ts
@@ -1,4 +1,4 @@
-import { ValueObject } from "@/contexts/Shared/Domain/ValueObject";
+import { ValueObject } from "@/contexts/Shared/domain/ValueObject";
 
 interface LanguageProps {
   name: string;
