+++ src/contexts/CoreContext/domain/valueObjects/UserName.ts
@@ -1,4 +1,4 @@
-import { ValueObject } from "@/contexts/Shared/Domain/ValueObject";
+import { ValueObject } from "@/contexts/Shared/domain/ValueObject";
 
 interface UserNameProps {
   name: string;
