+++ src/contexts/Shared/application/services/ResponseService.ts
@@ -1,5 +1,5 @@
 import { Response } from "express";
-import { IResponseEntity } from "../../Domain/interfaces/IResponseEntity";
+import { IResponseEntity } from "../../domain/interfaces/IResponseEntity";
 
 export class ResponseService {
   public static send(res: Response, response: IResponseEntity): void {
