+++ src/contexts/Shared/infrastructure/middlewares/ErrorHandlerMiddleware.ts
@@ -1,7 +1,7 @@
 import { NextFunction, Request, Response } from "express";
-import { ApiError } from "../errors/ApiError";
-import { ErrorResponseEntity } from "../../Domain/entity/ErrorResponseEntity";
 import { ResponseService } from "../../application/services/ResponseService";
+import { ErrorResponseEntity } from "../../domain/entity/ErrorResponseEntity";
+import { ApiError } from "../errors/ApiError";
 
 export class ErrorHandlerMiddleware {
   public static handle(
